pyproject.toml
[project]
name = "fastapi-genai-boilerplate"
version = "0.0.1"
description = "A production-ready FastAPI boilerplate designed for building scalable, AI-powered applications."
authors = [
  { name = "Keval Dekivadiya", email = "kevalkumar.dekivadiya@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10,<4.0"

dependencies = [
  "fastapi>=0.115.12,<0.116.0",
  "uvicorn>=0.34.3,<0.35.0",
  "loguru>=0.7.3,<0.8.0",
  "pydantic-settings>=2.9.1,<3.0.0",
  "fastapi-utils>=0.8.0,<0.9.0",
  "typing-inspect>=0.9.0,<0.10.0",
  "email-validator==2.2.0",
  "gunicorn>=23.0.0,<24.0.0",
  "pytest>=8.4.0",
  "setuptools<81",
  "prometheus-fastapi-instrumentator>=7.1.0,<8.0.0",
  "aiocache>=0.12.3,<1.0.0",
  "httpx>=0.28.1,<0.29.0",
  "celery[redis]>=5.5.3",
  "duckduckgo-search>=8.0.4",
  "langchain-community>=0.3.26",
  "langgraph>=0.4.9",
  "langchain-tavily>=0.2.4",
  "fastapi-limiter>=0.1.6",
  "fakeredis[lua]>=2.30.1",
  "langchain-openai>=0.3.26",
  "langfuse>=3.0.8",
  # RAG & Vector Database
  "google-genai>=1.0.0",
  "qdrant-client>=1.7.0",
  # PDF Processing
  "pypdf>=4.0.0",
  "pdfplumber>=0.11.0",
  "python-multipart>=0.0.9",
  # Database
  "sqlalchemy>=2.0.0",
  "asyncpg>=0.29.0",
  "alembic>=1.13.0",
]

[tool.setuptools]
packages = ["app"]

[dependency-groups]
dev = [
    "isort>=6.0.1",
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "coverage>=7.3",
    "black>=23.0",
    "mypy>=1.8",
    "pre-commit>=4.3.0",
    "ruff>=0.12.11",
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"
fix = true

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = "gunicorn.*"
ignore_missing_imports = true

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"
